change variable type

+++ convertor.py
@@ -1,5 +1,5 @@
 
 print("How many kilometers did you run today?")
 x = input()
-y = int(x) * round(0.621371192,3)
+y = float(x) * round(0.621371192,3)
 print(f"You run {y} miles today.")
